:truck: Rename inputProvider directory to input-provider
