+++ source/input-provider/AlbumDOMBasedInputProvider.js
