+++ source/input-provider/AlbumPromptBasedProvider.js
