+++ source/input-provider/StickerAlbumInputProvider.js
